requirements.txt
playwright==1.42.0
schedule==1.2.1
python-dotenv==1.0.1
Flask==3.0.2
Flask-Cors==4.0.0
